+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: `TextInputWithUnit`: Update styles.
 - Fix: `CrowdfundingCard`: Fix progress when value equals 0.
 
 ## [3.17.3] - 2021-06-07